# 🐛 Fixed hop energy calculation from neutral to positive SiDB