--- a/test/algorithms/simulation/sidb/quickexact.cpp
+++ b/test/algorithms/simulation/sidb/quickexact.cpp
@@ -928,27 +928,16 @@ TEMPLATE_TEST_CASE(
 
     const auto simulation_results = quickexact<TestType>(lyt, params);
 
-    REQUIRE(simulation_results.charge_distributions.size() == 3);
+    REQUIRE(simulation_results.charge_distributions.size() == 4);
 
-    auto energy_min = std::numeric_limits<double>::max();
-    for (const auto& layout : simulation_results.charge_distributions)
-    {
-        if (layout.get_system_energy() < energy_min)
-        {
-            energy_min = layout.get_system_energy();
-        }
-    }
+    const auto ground_state = std::min_element(
+        simulation_results.charge_distributions.cbegin(), simulation_results.charge_distributions.cend(),
+        [](const auto& lhs, const auto& rhs) { return lhs.get_system_energy() < rhs.get_system_energy(); });
 
-    for (const auto& layout : simulation_results.charge_distributions)
-    {
-        if (std::abs(layout.get_system_energy() - energy_min) < physical_constants::POP_STABILITY_ERR)
-        {
-            CHECK(layout.get_charge_state({1, 3, 0}) == sidb_charge_state::NEGATIVE);
-            CHECK(layout.get_charge_state({1, 3, 0}) == sidb_charge_state::NEGATIVE);
-            CHECK(layout.get_charge_state({2, 3, 0}) == sidb_charge_state::POSITIVE);
-            CHECK(layout.get_charge_state({3, 3, 0}) == sidb_charge_state::NEGATIVE);
-        }
-    }
+    CHECK(ground_state->get_charge_state({-1, 3, 0}) == sidb_charge_state::NEGATIVE);
+    CHECK(ground_state->get_charge_state({1, 3, 0}) == sidb_charge_state::POSITIVE);
+    CHECK(ground_state->get_charge_state({2, 3, 0}) == sidb_charge_state::NEGATIVE);
+    CHECK(ground_state->get_charge_state({3, 3, 0}) == sidb_charge_state::NEUTRAL);
 }
 
 TEMPLATE_TEST_CASE(
@@ -991,7 +980,7 @@ TEMPLATE_TEST_CASE(
 
     const auto simulation_results = quickexact<TestType>(lyt, params);
 
-    REQUIRE(simulation_results.charge_distributions.size() == 2);
+    REQUIRE(simulation_results.charge_distributions.size() == 4);
     const auto& charge_lyt_first = simulation_results.charge_distributions.front();
     CHECK_THAT(charge_lyt_first.get_system_energy(),
                Catch::Matchers::WithinAbs(0, physical_constants::POP_STABILITY_ERR));
@@ -1016,7 +1005,7 @@ TEMPLATE_TEST_CASE(
 
     const auto simulation_results = quickexact<TestType>(lyt, params);
 
-    REQUIRE(simulation_results.charge_distributions.size() == 3);
+    REQUIRE(simulation_results.charge_distributions.size() == 10);
     const auto& charge_lyt_first = simulation_results.charge_distributions.front();
     CHECK(charge_lyt_first.get_system_energy() < 0.08);
     CHECK(charge_lyt_first.get_system_energy() > -2.74);
@@ -1043,7 +1032,7 @@ TEMPLATE_TEST_CASE(
 
     const auto simulation_results = quickexact<TestType>(lyt, params);
 
-    CHECK(simulation_results.charge_distributions.size() == 5);
+    CHECK(simulation_results.charge_distributions.size() == 17);
 }
 
 TEMPLATE_TEST_CASE(
@@ -1109,7 +1098,26 @@ TEMPLATE_TEST_CASE(
 
     CHECK(lyt.num_cells() == 6);
 
-    CHECK(simulation_results.charge_distributions.size() == 1);
+    CHECK(simulation_results.charge_distributions.size() == 3);
+}
+
+TEMPLATE_TEST_CASE(
+    "4 DBs which can be only simulated if electron hop from negative to positiv is prohibited", "[quickexact]",
+    (cell_level_layout<sidb_technology, clocked_layout<cartesian_layout<siqad::coord_t>>>),
+    (charge_distribution_surface<cell_level_layout<sidb_technology, clocked_layout<cartesian_layout<siqad::coord_t>>>>))
+{
+    TestType lyt{};
+
+    lyt.assign_cell_type({0, 0, 1}, TestType::cell_type::NORMAL);
+    lyt.assign_cell_type({0, 0, 0}, TestType::cell_type::NORMAL);
+    lyt.assign_cell_type({3, 0, 1}, TestType::cell_type::NORMAL);
+    lyt.assign_cell_type({5, 0, 1}, TestType::cell_type::NORMAL);
+
+    const quickexact_params<TestType> params{sidb_simulation_parameters{3, -0.25}};
+
+    const auto simulation_results = quickexact<TestType>(lyt, params);
+
+    CHECK(simulation_results.charge_distributions.size() > 0);
 }
 
 TEMPLATE_TEST_CASE(
@@ -1237,6 +1245,22 @@ TEMPLATE_TEST_CASE(
         }
         CHECK(ground_state.size() == 1);
         CHECK(charge_index.size() == 1);
+    }
+
+    SECTION(
+        "Add SiDBs which are positively charged in the ground state, layout does not work fulfill the logic anymore.")
+    {
+        params.physical_parameters.base = 3;
+        lyt.assign_cell_type({15, 2, 1}, TestType::cell_type::NORMAL);
+        lyt.assign_cell_type({15, 2, 0}, TestType::cell_type::NORMAL);
+
+        const auto simulation_results = quickexact<TestType>(lyt, params);
+        // find the ground state, which is the charge distribution with the lowest energy
+        const auto ground_state = std::min_element(
+            simulation_results.charge_distributions.cbegin(), simulation_results.charge_distributions.cend(),
+            [](const auto& lhs, const auto& rhs) { return lhs.get_system_energy() < rhs.get_system_energy(); });
+
+        CHECK(ground_state->num_positive_sidbs() > 0);
     }
 
     SECTION("Standard Physical Parameters")
@@ -1407,6 +1431,10 @@ TEMPLATE_TEST_CASE(
         const auto ground_state = std::min_element(
             simulation_results.charge_distributions.cbegin(), simulation_results.charge_distributions.cend(),
             [](const auto& lhs, const auto& rhs) { return lhs.get_system_energy() < rhs.get_system_energy(); });
+
+        CHECK(ground_state->num_negative_sidbs() == 5);
+        CHECK(ground_state->num_neutral_sidbs() == 4);
+        CHECK(ground_state->num_positive_sidbs() == 0);
 
         // check that charge distribution is correct; binary 1 is propagated through the BDL wire
         CHECK(ground_state->get_charge_state({0, 0, 0}) == sidb_charge_state::NEGATIVE);