## Description:

This PR fixes a major bug. The hopping energy from a neutrally to a positively charged SiDB was calculated incorrectly. Finally, this fix allows the successful simulation of positively charged SiDBs.

@samuelngsh I would strongly recommend to make this change in *SiQAD* as well. Many thanks! 

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [ ] I have made sure that all CI jobs on GitHub pass.
- [ ] The pull request introduces no new warnings and follows the project's style guidelines.