@Drewniok 2024-01-10
Note: The changes made lead to slightly better benchmarking results.

``main:``
```
benchmark name                       samples       iterations    est run time
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
QuickExact                                     100             1     1.69999 m 
                                          1.0291 s      1.0275 s     1.03124 s 
                                        9.34506 ms    7.36008 ms    14.3194 ms 
                                                                               
QuickSim                                       100             1    422.327 ms 
                                        4.29704 ms    4.26032 ms    4.34201 ms 
                                        206.767 us    171.362 us    260.536 us 
```

``PR #364:``
```
benchmark name                       samples       iterations    est run time
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
QuickExact                                     100             1     1.69137 m 
                                         1.01837 s     1.01691 s     1.02041 s 
                                        8.72947 ms    6.78313 ms    11.2649 ms 
                                                                               
QuickSim                                       100             1     447.09 ms 
                                        4.47946 ms    4.44835 ms     4.5125 ms 
                                         163.73 us    145.971 us    189.081 us 
```